api/requirements.txt
Flask
sqlalchemy
google-cloud-alloydb-connector[pg8000]
pandas-gbq
google_auth_oauthlib